-- consump log  model

+++ backend/src/models/consumption_log.rs
@@ -1,3 +1,4 @@
+use bigdecimal::BigDecimal;
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use sqlx::prelude::FromRow;
@@ -8,9 +9,9 @@ use uuid::Uuid;
 pub struct ConsumptionLog {
     pub id:                        Uuid,
     pub device_id:                 Uuid,
-    pub units_remaining:           f64,
-    pub previous_reading:          Option<f64>,
-    pub consumption_rate:          Option<f64>,
+    pub units_remaining:           BigDecimal,
+    pub previous_reading:          Option<BigDecimal>,
+    pub consumption_rate:          Option<BigDecimal>,
     pub reading_trigger:           String,
     pub recorded_at:               DateTime<Utc>
 }
